Fully functional GPIO microapp interrupts
The GPIO interrupts arrive and can be used to initiate microapp
behaviour. It is also possible to perform Serial.write commands within
the interrupt.

+++ include/Arduino.h
@@ -25,7 +25,7 @@ void loop();
 //
 // A delay in ms. Hence 1000 means a delay of one second.
 //
-void delay(uint16_t delay_ms);
+void delay(uint32_t delay_ms);
 
 //
 // A bunch of functions that are mainly useful in either development mode, on a Crownstone that is embedded
